Bugfix, rsaN to BigInteger

+++ src/main/java/com/master_thesis/client/RSAThreshold.java
@@ -42,7 +42,7 @@ public class RSAThreshold extends HomomorphicHash {
         shares.forEach((uri, secretShare) -> {
             secretShare.setMatrixOfClient(matrixOfClient);
             secretShare.setSkShare(skShares);
-            secretShare.setRsaN(rsaN.intValue());
+            secretShare.setRsaN(rsaN);
             secretShare.setPublicKey(publicKey.intValue());
         });
